pyproject.toml
[project]
name = "phd-thesis-fitting"
version = "0.1.0"
description = "Notebooks for fitting the wear coefficient from measurements."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jupyter",
    "jupytext",
    'pyroll-core',
    "pyroll-report",
    "pyroll-ring-model-thermal",
    "pyroll-freiberg-flow-stress",
    "pyroll-hensel-force-torque",
    "pyroll-pillar-model",
    "pyroll-neutral-point-estimator",
    "pyroll-sparling-spreading",
    "lmfit",
    "matplotlib",
    "pandas",
    "scikit-learn",
    "numdifftools"
]

[dependency-groups]
dev = [
    "pre-commit~=4.1.0",
]
